backend/requirements-core.txt
# JanVedha AI — Core Backend Dependencies (Lightweight)
# Excludes heavy ML/AI/Scraping libraries for faster setup

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi
uvicorn[standard]
pydantic-settings
pydantic[email]
python-multipart

# ── MongoDB ───────────────────────────────────────────────────────────────────
motor>=3.3.0
beanie>=1.25.0

# ── Auth & Security ───────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]
PyJWT==2.8.0

# ── Caching / Sessions ────────────────────────────────────────────────────────
redis==5.0.3

# ── Storage ───────────────────────────────────────────────────────────────────
aiofiles

# ── HTTP ──────────────────────────────────────────────────────────────────────
httpx
python-telegram-bot[webhooks]>=21.0
google-cloud-speech
google-cloud-texttospeech

# ── Documents / Templates (Optional but lightweight) ──────────────────────────
jinja2

# ── RAG and Observability ─────────────────────────────────────────────────────
chromadb>=0.4.24
sentence-transformers>=2.3.1
rank-bm25>=0.2.2
tiktoken>=0.6.0
langfuse>=2.9.0
PyMuPDF>=1.23.21
